package.json and fixed schema file with avg_rating fixed

+++ db/schema.sql
@@ -7,7 +7,7 @@ CREATE TABLE `gameinfo` (
     game_name VARCHAR(255) NOT NULL,
     instructions VARCHAR(255) NOT NULL,
     category VARCHAR(255) NOT NULL,
-    avg_rating FLOAT(255) DEFAULT NULL,
+    avg_rating FLOAT(5,4) DEFAULT NULL,
    	min_players TINYINT NOT NULL,
    	max_players TINYINT DEFAULT NULL,
   	date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
